+++ R/plot_annual_cumulative_stats.R
@@ -165,7 +165,7 @@ plot_annual_cumulative_stats <- function(data,
         ggplot2::xlab("Year")+
         ggplot2::scale_color_brewer(palette = "Set1") +
         ggplot2::theme_bw() +
-        ggplot2::guides(colour = FALSE)+
+        ggplot2::guides(colour = 'none') +
         {if (include_title & .y != "XXXXXXX") ggplot2::ggtitle(paste(.y)) } +
         ggplot2::theme(panel.border = ggplot2::element_rect(colour = "black", fill = NA, size = 1),
                        panel.grid = ggplot2::element_line(size = .2),
@@ -203,7 +203,7 @@ plot_annual_cumulative_stats <- function(data,
           ggplot2::xlab("Year")+
           ggplot2::scale_color_brewer(palette = "Set1") +
           ggplot2::theme_bw() +
-          ggplot2::guides(colour = FALSE) +
+          ggplot2::guides(colour = 'none') +
           {if (include_title & .y != "XXXXXXX") ggplot2::ggtitle(paste(.y)) } +
           ggplot2::theme(panel.border = ggplot2::element_rect(colour = "black", fill = NA, size = 1),
                          panel.grid = ggplot2::element_line(size = .2),